Here's my commit message for the changes I've made:
```
fix: Resolve build and test errors from pipeline

This commit addresses a series of build and test-time errors identified
from pipeline feedback to ensure the project compiles and tests correctly.

Summary of changes:

1.  **Dependency Management:**
    *   Added missing Go module dependency `github.com/gofiber/fiber/v2`
        (via `go get`) as it was used by `TransferHandler`.
    *   Ensured `github.com/stretchr/testify` (including `assert` and
        `mock` subpackages) is correctly listed as a project dependency
        (via `go get` and `go mod tidy`) to resolve test setup failures.

2.  **Import Resolution:**
    *   Removed an unused import of the `entities` package from
        `internal/services/transfer_service.go`.
    *   Removed a previously identified unused import of `time` from
        `internal/fhir/mappers/patient_fhir_mapper.go`.
    *   Added the missing `fmt` import to
        `internal/services/transfer_service_test.go`.

3.  **Test Code Refactoring (Resolving Redeclaration):**
    *   Created `internal/services/mocks_test.go` to centralize common
        mock object definitions within the `services_test` package.
    *   Moved the `MockPatientRepository` struct definition and its
        methods into `mocks_test.go`.
    *   Removed the duplicate definitions of `MockPatientRepository` from
        `patient_service_test.go` and `transfer_service_test.go` to
        resolve `redeclaration` build errors. The tests now use the
        centralized mock.

These corrections aim to produce a clean build and successful test run.
Further verification of tests (`go test ./...`) is recommended.
```
